Using the Writer in my Specs
